+++ test/bigtest/network/factories/feefineactions.js
@@ -5,7 +5,7 @@ export default Factory.extend({
   accountId: '8fd350aa-f6f5-44ec-ace3-0b96d72d20e5',
   amountAction: 10,
   balance: 10,
-  comments: '',
+  comments: (i) => 'This is a comment' + i,
   createdAt: 'Main Circ',
   dateAction: '2019-04-22T22:07:07.343+0000',
   source: 'ADMINISTRATOR, DIKU',